auto-destroy wearables if placed back on parent

+++ DomainContent/AvatarStore/attachmentItemScript.js
@@ -240,13 +240,10 @@
                 }
             }
             
-            if (!isAttached) { 
-                // If placed back within NOT_ATTACHED_DESTROY_RADIUS of the original parent entity 
-                // and it is not attached then destroy it (i.e. user is putting it back)
-                if (Vec3.distance(initialParentPosition, position) < NOT_ATTACHED_DESTROY_RADIUS) {
-                    Entities.deleteEntity(_entityID);
-                    return;
-                }
+            // If placed back within NOT_ATTACHED_DESTROY_RADIUS of the original parent entity 
+            // and it is not attached then destroy it (i.e. user is putting it back)
+            if (!isAttached && Vec3.distance(initialParentPosition, position) < NOT_ATTACHED_DESTROY_RADIUS) {
+                Entities.deleteEntity(_entityID);
             }
         }
     };
